+++ autonav_ws/src/autonav_pathing/src/object_detection/circumscription_test.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 
-class circumscription_tester:
+class Circumscription_Tester:
     def initialize(self):
         self.blank = None
