+++ path_complete/path_planning/PlanningTest.py
@@ -1,30 +1,38 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import math
-import path_planning
+import TangentBased
 import time
 
 def planning_test(wpoints, obstacles):
-    test = path_planning.path_planning()
-    test.setpath(wpoints)
-    test.setobstacles(obstacles)
+    testcw = TangentBased.path_planning()
+    testcw.setpath(wpoints)
+    testcw.setobstacles(obstacles)
     
+    testccw = TangentBased.path_planning()
+    testccw.setpath(wpoints)
+    testccw.setobstacles(obstacles)
 
     # run test.intersections until no new points are added
     
     start_time = time.time()
     
     for i in range(10):
-        print("intersections called")
-        test.intersections("cw")
-        test.path_intersections()
+        #print("intersections called")
+        testcw.intersections("cw")
+        testcw.path_intersections()
+        testcw.delete_inside()
+
+        testccw.intersections("ccw")
+        testccw.path_intersections()
+        testccw.delete_inside()
     
     execution_time = time.time() - start_time
     
     
 
     # here
-    print(f"Start test, test.final {test.final}")
+    print(f"Start test, test.final {testcw.final}")
     obstacle_circles = []
     obstacle_circles_2 = []
     for i in range(len(obstacles)):
@@ -44,8 +52,8 @@ def planning_test(wpoints, obstacles):
         ax2.add_patch(obstacle_circles[i])
 
     original_path = []
-    for i in range(len(test.path)):
-        original_path.append(test.path[i][0:2])
+    for i in range(len(testcw.path)):
+        original_path.append(testcw.path[i][0:2])
 
     ex, why = zip(*original_path)
     print(f"original_path {original_path}")
@@ -66,9 +74,9 @@ def planning_test(wpoints, obstacles):
     final_path = []
 
 
-    for i in range(len(test.final)):
+    for i in range(len(testcw.final)):
         #print(f"test.final[i] {test.final[i]}")
-        final_path.append(test.final[i][0:2])
+        final_path.append(testcw.final[i][0:2])
     #print(f"final path in PATH PLANNING TESTER {final_path}")
     x, y = zip(*final_path)
